# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XF Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
}

export interface AuthResponse {
  user: User;
}

export interface SignupBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface OrderItem {
  productId: string;
  name: string;
  price: number;
  size: string;
  quantity: number;
}

export interface ShippingAddress {
  street: string;
  city: string;
  zip: string;
  country: string;
}

export interface Order {
  id: number;
  /** @nullable */
  userId?: number | null;
  customerName: string;
  email: string;
  shippingAddress: ShippingAddress;
  items: OrderItem[];
  totalPrice: number;
  status: string;
  createdAt: string;
}

export interface CreateOrderBody {
  customerName: string;
  email: string;
  shippingAddress: ShippingAddress;
  items: OrderItem[];
}

export interface UpdateOrderStatusBody {
  status: string;
}