# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XF Store API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new account
 */
export const SignupBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
});

export const SignupResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    role: zod.string(),
  }),
});

/**
 * @summary Log in with email and password
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    role: zod.string(),
  }),
});

/**
 * @summary Log out current session
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current logged-in user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  role: zod.string(),
});

/**
 * @summary List current user's orders
 */
export const ListMyOrdersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number().nullish(),
  customerName: zod.string(),
  email: zod.string(),
  shippingAddress: zod.object({
    street: zod.string(),
    city: zod.string(),
    zip: zod.string(),
    country: zod.string(),
  }),
  items: zod.array(
    zod.object({
      productId: zod.string(),
      name: zod.string(),
      price: zod.number(),
      size: zod.string(),
      quantity: zod.number(),
    }),
  ),
  totalPrice: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const ListMyOrdersResponse = zod.array(ListMyOrdersResponseItem);

/**
 * @summary Place a new order
 */
export const CreateOrderBody = zod.object({
  customerName: zod.string(),
  email: zod.string(),
  shippingAddress: zod.object({
    street: zod.string(),
    city: zod.string(),
    zip: zod.string(),
    country: zod.string(),
  }),
  items: zod.array(
    zod.object({
      productId: zod.string(),
      name: zod.string(),
      price: zod.number(),
      size: zod.string(),
      quantity: zod.number(),
    }),
  ),
});

/**
 * @summary List all orders (admin only)
 */
export const AdminListOrdersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number().nullish(),
  customerName: zod.string(),
  email: zod.string(),
  shippingAddress: zod.object({
    street: zod.string(),
    city: zod.string(),
    zip: zod.string(),
    country: zod.string(),
  }),
  items: zod.array(
    zod.object({
      productId: zod.string(),
      name: zod.string(),
      price: zod.number(),
      size: zod.string(),
      quantity: zod.number(),
    }),
  ),
  totalPrice: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});
export const AdminListOrdersResponse = zod.array(AdminListOrdersResponseItem);

/**
 * @summary Update order status (admin only)
 */
export const AdminUpdateOrderStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminUpdateOrderStatusBody = zod.object({
  status: zod.string(),
});

export const AdminUpdateOrderStatusResponse = zod.object({
  id: zod.number(),
  userId: zod.number().nullish(),
  customerName: zod.string(),
  email: zod.string(),
  shippingAddress: zod.object({
    street: zod.string(),
    city: zod.string(),
    zip: zod.string(),
    country: zod.string(),
  }),
  items: zod.array(
    zod.object({
      productId: zod.string(),
      name: zod.string(),
      price: zod.number(),
      size: zod.string(),
      quantity: zod.number(),
    }),
  ),
  totalPrice: zod.number(),
  status: zod.string(),
  createdAt: zod.string(),
});